requirements.txt
fastapi
requests
openai
google-generativeai

